refactor: refine code for pr 1958
